requirements.txt
aniso8601==9.0.1
APScheduler==3.11.0
astroid==3.3.5
black==24.10.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dill==0.3.9
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.3
pycodestyle==2.12.1
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.35
tomlkit==0.13.2
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.4
